requirements.txt
aws-cdk-lib==2.97.*
constructs>=10.0.0,<11.0.0
